backend/requirements.txt
fastapi==0.115.7
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
pymysql==1.1.1
python-dotenv==1.0.1
email-validator==2.2.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
